(PC-10896) Move authentication functions to core.users.api
